:tuck: Rename create-user
